+++ unit-1/list application/polynomial/term.h
@@ -3,14 +3,14 @@
 
 struct term
 {
-	int coeff_;
-	int expo_;
+	float coeff_;
+	float expo_;
 };
 typedef struct term term_t;
 
 void set_term(term_t* ptr_term, int coeff, int expo_);
 void disp_term(term_t* ptr_term);
 int compare_exponents(term_t* left, term_t* right);
-int term_value(term_t* ptr_term, int x);
+float term_value(term_t* ptr_term, int x);
 
 #endif
